JS-5228: Move chat form to separate component
